requirements-test.txt
-r requirements.txt
colorclass
docopt
pytest-cov
requests
